requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
beanie==1.13.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-users==10.2.1
fastapi-users-db-beanie==1.1.2
fastapi-users-db-sqlalchemy==4.0.3
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
makefun==1.15.0
motor==3.1.1
multidict==6.0.3
optional==0.0.1
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
